+++ TRENZ.Extensions.Infisical/InfisicalSecretsRepository.cs
@@ -89,24 +89,33 @@ public class InfisicalSecretsRepository(
             {
                 return client.ListSecrets(request).ToFrozenDictionary(s => s.SecretKey);
             }
-            catch (InfisicalException e)
+            catch (InfisicalException e) 
+                when (e.InnerException != null && 
+                      e.InnerException.Message.Contains("Error during GET request: Unexpected response: Unauthorized"))
             {
-                logger?.LogInformation(e, "Failed to load secrets");
+                logger?.LogError(e, "Failed to load secrets: check credentials to Infisical secrets instance.");
 
+                return null;
+            }
+            catch (Exception e)
+            {
                 retries++;
+                
+                logger?.LogWarning(e, "Failed to load secrets, attempt #{Try}", retries);
+
                 if (retries >= maxRetries)
                 {
                     logger?.LogCritical(e, "Max retries exceeded");
 
                     return null;
                 }
-
-                // can't use async here, see https://github.com/dotnet/runtime/issues/36018
-
-                // back off exponentially but at least 50ms
-                var backoff = 50 + 5 * Math.Pow(2, retries);
-                Thread.Sleep((int)backoff);
             }
+
+            // can't use async here, see https://github.com/dotnet/runtime/issues/36018
+
+            // back off exponentially but at least 50ms
+            var backoff = 50 + 5 * Math.Pow(2, retries);
+            Thread.Sleep((int)backoff);            
         }
     }
 }