+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- In case loading secrets fails, the log message now contains the number of retries
+
 # 1.1.2
 
 - Bump release