+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - In case loading secrets fails, the log message now contains the number of retries
 - When authorization fails, we stop retrying and log the failure
 - Omitting the project ID will now throw
+- Upgrade the Infisical SDK package to 3.0.4
 
 # 1.1.2